# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Oqu REST API",
        "contact": {}
    },
    "paths": {
        "/admin/courses": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "description": "New course",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewCourse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course with id 1"
                    },
                    "400": {
                        "description": "Incorrect request body"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [admin] can access! Your role is [user or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/courses/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Course"
                        }
                    },
                    "400": {
                        "description": "Provide number"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [admin] can access! Your role is [user or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/courses/{id}/lessons": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "description": "New lesson",
                        "name": "lesson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewLesson"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson with id 1"
                    },
                    "400": {
                        "description": "Incorrect request body"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [admin] can access! Your role is [user or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/lessons/{id}/test": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AdminTestView"
                        }
                    },
                    "400": {
                        "description": "Provide number"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [admin] can access! Your role is [user or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "description": "New test",
                        "name": "test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewTest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Lesson id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New test"
                    },
                    "400": {
                        "description": "Json error"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [admin] can access! Your role is [user or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "400": {
                        "description": "Provide number"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [admin] can access! Your role is [user or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [admin] can access! Your role is [user or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AdminTestView": {
            "type": "object",
            "properties": {
                "answer_options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Answer"
                    }
                },
                "question": {
                    "type": "string"
                },
                "question_id": {
                    "type": "integer"
                }
            }
        },
        "models.Answer": {
            "type": "object",
            "properties": {
                "is_correct": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.Course": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.NewCourse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.NewLesson": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.NewTest": {
            "type": "object",
            "properties": {
                "answer_options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Answer"
                    }
                },
                "question": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}